migrated router logic to controller
